Add game_of_life demo to the showcase
